Use local docker images
By 'eval $(minikube docker-env)', the docker commands are redirected to be
executed in the minikube environment and the images are also stored there.
Adapting the yml files then leads to the fact that these are used.

Rebuilding the jar backend after commenting out the crash in 'fail()' is needed
to use a updated jar for the v0.2 of the todobackend image.